refactor : menu -> domain 으로 디렉토리명 변경
