+++ src/app/handler/implement/OrderCancelHandler.java
@@ -2,7 +2,7 @@ package app.handler.implement;
 
 import app.command.implement.OrderCancelCommand;
 import app.handler.Handler;
-import app.menu.Category;
+import app.domain.Category;
 
 public class OrderCancelHandler extends Handler {
